refactor: ReviewComment에서 사용하는 기능을 hook으로 분리
